+++ CHANGELOG.md
@@ -47,6 +47,7 @@ the detailed section referring to by linking pull requests or issues.
 * Support for HTTP-based provisioning (#963)
 * Let Control Plane delegate data transfer to Data Plane (#988)
 * CosmosDb based `PolicyStore` (#826)
+* Http Provisioner Webhook endpoint (#1039)
 
 #### Changed